+++ R/daemon.R
@@ -141,7 +141,7 @@ daemon <- function(url, autoexit = TRUE, cleanup = TRUE, output = FALSE,
     ctx <- .context(sock)
     aio <- recv_aio(ctx, mode = 1L, timeout = idletime, cv = cv)
     wait(cv) || break
-    m <- .subset2(aio, "data")
+    m <- collect_aio(aio)
     is.environment(m) || {
       count < timerstart && {
         start <- mclock()
@@ -175,18 +175,20 @@ daemon <- function(url, autoexit = TRUE, cleanup = TRUE, output = FALSE,
 #'
 #' @inheritParams daemon
 #'
-#' @return Invisible NULL.
+#' @return Logical TRUE or FALSE.
 #'
 #' @keywords internal
 #' @export
 #'
 .daemon <- function(url) {
 
-  sock <- socket(protocol = "rep", dial = url, autostart = NA)
+  cv <- cv()
+  sock <- socket(protocol = "rep")
   on.exit(reap(sock))
+  pipe_notify(sock, cv = cv, remove = TRUE)
+  dial(sock, url = url, autostart = NA, error = TRUE)
   data <- eval_mirai(recv(sock, mode = 1L, block = TRUE))
-  send(sock, data = data, mode = 1L, block = TRUE)
-  msleep(2000L)
+  send(sock, data = data, mode = 1L) || until(cv, .limit_short)
 
 }