+++ R/mirai-package.R
@@ -16,7 +16,7 @@
 
 #' mirai: Minimalist Async Evaluation Framework for R
 #'
-#' High performance parallel code execution and distributed computing. Designed
+#' High-performance parallel code execution and distributed computing. Designed
 #'     for simplicity, a 'mirai' evaluates an R expression asynchronously, on
 #'     local or network resources, resolving automatically upon completion.
 #'     Modern networking and concurrency built on 'nanonext' and 'NNG' (Nanomsg
@@ -85,6 +85,7 @@
 ._ <- list2env(
   list(
     arglen = "'args' and/or 'url' must be of length 1 or the same length",
+    character_class = "'class' must be a character string",
     cluster_inactive = "cluster is no longer active",
     correct_context = "'host' must be specified if not using directly in a function argument",
     daemons_unset = "a numeric value for 'url' requires daemons to be set",
@@ -98,9 +99,9 @@
     register_cluster = "this function requires a more recent version of R",
     requires_daemons = "launching one local daemon as none previously set",
     requires_local = "SSH tunnelling requires 'url' hostname to be '127.0.0.1' or 'localhost'",
-    refhook_invalid = "'refhook' must be a list of 2 functions or NULL",
+    serial_invalid = "'fns' must be a list of 2 functions or NULL",
     single_url = "only one 'url' should be specified",
-    sync_timeout = "initial sync with dispatcher timed out after 10s",
+    sync_timeout = "initial sync with dispatcher/daemon timed out after 10s",
     url_spec = "numeric value for 'url' is out of bounds",
     wrong_dots = "'...' arguments should only be of integer, numeric or logical type"
   ),